Move header and footer to App, setup basic routing
