+++ theme/package.json
@@ -18,6 +18,7 @@
     "@apollo/client": "^3.4.8",
     "@vue/apollo-composable": "^4.0.0-alpha.14",
     "graphql": "^15.5.1",
-    "graphql-tag": "^2.12.5"
+    "graphql-tag": "^2.12.5",
+    "vue-router": "^4.0.11"
   }
 }
